Change background color of sidebar blocks from gray to white
Closes Vulkan-Site#56

Note, this affects *all* sidebar blocks in the document, not just the
Valid Usage blocks. There are a lot of sidebar blocks in the spec
content besides VU blocks, and it will also affect the other components.
I don't think this is a drawback - if it's hard to read for VU
statements, it's hard to read for anything else as well.